feat: 피드 좋아요 기능 추가 (#46)
* feat: 피드 좋아요 엔티티인 FeedLike 추가

* feat: 피드에 좋아요 버튼을 누를 경우 반환할 FeedLikeResponseDto 추가

* feat: Feed 엔티티의 likeCount를 int에서 long으로 변경 및 likeCount 증감 메서드 추가

* feat: 피드 좋아요 토글 기능 추가

* refactor: feed 테이블의 like_count 타입을 BIGINT로 변경

* feat: 피드 좋아요 동시성 문제 해결을 위해 좋아요 수 증감은 DB레벨의 원자적 업데이트 사용하도록 변경

* fix: 피드 좋아요 관련 영속성 컨텍스트와 DB 불일치 해결

* feat: 배포, 개발 DB에 피드좋아요 테이블 추가

* feat: 피드에 좋아요를 누른 사용자 목록 조회 기능 추가

* feat: Swagger 적용 및 로그인 어노테이션 적용

+++ src/main/java/com/starterpack/feed/dto/FeedResponseDto.java
@@ -14,7 +14,7 @@ public record FeedResponseDto(
         String imageUrl,
         FeedType feedType,
         CategoryResponseDto category,
-        int likeCount,
+        Long likeCount,
         LocalDateTime createdAt,
 
         @JsonInclude(JsonInclude.Include.NON_EMPTY)
